Fixing tests for release/2.7.1
