+++ cypress/e2e/00_quick/03_ignore_pref_spec.js
@@ -21,7 +21,6 @@ describe('03 Ignore tweets: ', function () {
         cy.get(".app-tweet-table", {timeout: 30000}).then(panel => {
             cy.wait(30000);
             // Make sure other tests have not hidden the tweet we're testing against
-            cy.unhideTweets(1);
             if (panel.find(POWYS_TWEET).length > 0) {
                 cy.get(POWYS_TWEET_VISIBLE, {timeout: 60000});
                 cy.get(POWYS_TWEET_VISIBLE, {timeout: 60000}).should('be.visible');
